+++ engine-spring/core/src/main/java/org/camunda/bpm/engine/spring/ProcessEngineFactoryBean.java
@@ -19,6 +19,7 @@ package org.camunda.bpm.engine.spring;
 import org.camunda.bpm.engine.ProcessEngine;
 import org.camunda.bpm.engine.impl.ProcessEngineImpl;
 import org.camunda.bpm.engine.impl.cfg.ProcessEngineConfigurationImpl;
+import org.camunda.bpm.engine.spring.impl.SpringExpressionManager;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.FactoryBean;
