+++ engine-spring/core/src/main/java/org/camunda/bpm/engine/spring/components/ActivitiContextUtils.java
@@ -25,11 +25,11 @@ import org.camunda.bpm.engine.spring.components.config.xml.StateHandlerAnnotatio
   */
 public class ActivitiContextUtils {
 
-	public static final String ANNOTATION_STATE_HANDLER_BEAN_FACTORY_POST_PROCESSOR_BEAN_NAME= StateHandlerAnnotationBeanFactoryPostProcessor.class.getName().toLowerCase();
-	/**
-	 * the name of the default registry used to store all state handling components
-	 */
-	public final static String ACTIVITI_REGISTRY_BEAN_NAME = "activitiComponentRegistry" ;
+  public static final String ANNOTATION_STATE_HANDLER_BEAN_FACTORY_POST_PROCESSOR_BEAN_NAME= StateHandlerAnnotationBeanFactoryPostProcessor.class.getName().toLowerCase();
+  /**
+   * the name of the default registry used to store all state handling components
+   */
+  public final static String ACTIVITI_REGISTRY_BEAN_NAME = "activitiComponentRegistry" ;
 
 
 }
