+++ engine-spring/core/src/main/java/org/camunda/bpm/engine/spring/components/aop/ProcessStartingMethodInterceptor.java
@@ -43,130 +43,130 @@ import org.springframework.util.StringUtils;
  */
 public class ProcessStartingMethodInterceptor implements MethodInterceptor {
 
-	private Logger log = Logger.getLogger(getClass().getName());
+  private Logger log = Logger.getLogger(getClass().getName());
 
-	/**
-	 * injected reference - can be obtained via a {@link org.camunda.bpm.engine.spring.ProcessEngineFactoryBean}
-	 */
-	protected ProcessEngine processEngine;
+  /**
+   * injected reference - can be obtained via a {@link org.camunda.bpm.engine.spring.ProcessEngineFactoryBean}
+   */
+  protected ProcessEngine processEngine;
 
-	/**
-	 * @param processEngine takes a reference to a {@link org.camunda.bpm.engine.ProcessEngine}
-	 */
-	public ProcessStartingMethodInterceptor(ProcessEngine processEngine) {
-		this.processEngine = processEngine;
-	}
+  /**
+   * @param processEngine takes a reference to a {@link org.camunda.bpm.engine.ProcessEngine}
+   */
+  public ProcessStartingMethodInterceptor(ProcessEngine processEngine) {
+    this.processEngine = processEngine;
+  }
 
-	boolean shouldReturnProcessInstance(StartProcess startProcess, MethodInvocation methodInvocation, Object result) {
-		return (result instanceof ProcessInstance || methodInvocation.getMethod().getReturnType().isAssignableFrom(ProcessInstance.class));
-	}
+  boolean shouldReturnProcessInstance(StartProcess startProcess, MethodInvocation methodInvocation, Object result) {
+    return (result instanceof ProcessInstance || methodInvocation.getMethod().getReturnType().isAssignableFrom(ProcessInstance.class));
+  }
 
-	boolean shouldReturnProcessInstanceId(StartProcess startProcess, MethodInvocation methodInvocation, Object result) {
-		return startProcess.returnProcessInstanceId() && (result instanceof String || methodInvocation.getMethod().getReturnType().isAssignableFrom(String.class));
-	}
+  boolean shouldReturnProcessInstanceId(StartProcess startProcess, MethodInvocation methodInvocation, Object result) {
+    return startProcess.returnProcessInstanceId() && (result instanceof String || methodInvocation.getMethod().getReturnType().isAssignableFrom(String.class));
+  }
 
-	@SuppressWarnings("unused")
-	boolean shouldReturnAsyncResultWithProcessInstance(StartProcess startProcess, MethodInvocation methodInvocation, Object result) {
-		return (result instanceof Future || methodInvocation.getMethod().getReturnType().isAssignableFrom(Future.class));
-	}
+  @SuppressWarnings("unused")
+  boolean shouldReturnAsyncResultWithProcessInstance(StartProcess startProcess, MethodInvocation methodInvocation, Object result) {
+    return (result instanceof Future || methodInvocation.getMethod().getReturnType().isAssignableFrom(Future.class));
+  }
 
-	public Object invoke(MethodInvocation invocation) throws Throwable {
+  public Object invoke(MethodInvocation invocation) throws Throwable {
 
-		Method method = invocation.getMethod();
+    Method method = invocation.getMethod();
 
-		StartProcess startProcess = AnnotationUtils.getAnnotation(method, StartProcess.class);
+    StartProcess startProcess = AnnotationUtils.getAnnotation(method, StartProcess.class);
 
-		String processKey = startProcess.processKey();
+    String processKey = startProcess.processKey();
 
-		Assert.hasText(processKey, "you must provide the name of process to start");
+    Assert.hasText(processKey, "you must provide the name of process to start");
 
-		Object result;
-		try {
-			result = invocation.proceed();
-			Map<String, Object> vars = this.processVariablesFromAnnotations(invocation);
+    Object result;
+    try {
+      result = invocation.proceed();
+      Map<String, Object> vars = this.processVariablesFromAnnotations(invocation);
 
-			String businessKey = this.processBusinessKey(invocation);
+      String businessKey = this.processBusinessKey(invocation);
 
-			log.info("variables for the started process: " + vars.toString());
+      log.info("variables for the started process: " + vars.toString());
 
-			RuntimeService runtimeService = this.processEngine.getRuntimeService();
-			ProcessInstance pi ;
-			if (null != businessKey && StringUtils.hasText(businessKey)) {
-				pi = runtimeService.startProcessInstanceByKey(processKey, businessKey, vars);
-				log.info("the business key for the started process is '" + businessKey + "' ");
-			} else {
-				pi = runtimeService.startProcessInstanceByKey(processKey, vars);
-			}
+      RuntimeService runtimeService = this.processEngine.getRuntimeService();
+      ProcessInstance pi ;
+      if (null != businessKey && StringUtils.hasText(businessKey)) {
+        pi = runtimeService.startProcessInstanceByKey(processKey, businessKey, vars);
+        log.info("the business key for the started process is '" + businessKey + "' ");
+      } else {
+        pi = runtimeService.startProcessInstanceByKey(processKey, vars);
+      }
 
-			String pId = pi.getId();
+      String pId = pi.getId();
 
-			if (invocation.getMethod().getReturnType().equals(void.class))
-				return null;
+      if (invocation.getMethod().getReturnType().equals(void.class))
+        return null;
 
-			if (shouldReturnProcessInstance(startProcess, invocation, result))
-				return pi;
+      if (shouldReturnProcessInstance(startProcess, invocation, result))
+        return pi;
 
-			if (shouldReturnProcessInstanceId(startProcess, invocation, result))
-				return pId;
+      if (shouldReturnProcessInstanceId(startProcess, invocation, result))
+        return pId;
 
-			if (shouldReturnAsyncResultWithProcessInstance(startProcess, invocation, result)) {
-				return new AsyncResult<ProcessInstance>(pi);
-			}
+      if (shouldReturnAsyncResultWithProcessInstance(startProcess, invocation, result)) {
+        return new AsyncResult<ProcessInstance>(pi);
+      }
 
-		} catch (Throwable th) {
-			throw new RuntimeException(th);
-		}
-		return result;
-	}
+    } catch (Throwable th) {
+      throw new RuntimeException(th);
+    }
+    return result;
+  }
 
-	protected String processBusinessKey(MethodInvocation invocation) throws Throwable {
-		Map<BusinessKey, String> businessKeyAnnotations = this.mapOfAnnotationValues( BusinessKey.class ,invocation);
-		if (businessKeyAnnotations.size() == 1) {
-			BusinessKey processId = businessKeyAnnotations.keySet().iterator().next();
-			return businessKeyAnnotations.get(processId);
-		}
-		return null;
-	}
+  protected String processBusinessKey(MethodInvocation invocation) throws Throwable {
+    Map<BusinessKey, String> businessKeyAnnotations = this.mapOfAnnotationValues( BusinessKey.class ,invocation);
+    if (businessKeyAnnotations.size() == 1) {
+      BusinessKey processId = businessKeyAnnotations.keySet().iterator().next();
+      return businessKeyAnnotations.get(processId);
+    }
+    return null;
+  }
 
-	@SuppressWarnings("unchecked")
-	private <K extends Annotation, V> Map<K, V> mapOfAnnotationValues(Class<K> annotationType, MethodInvocation invocation) {
-		Method method = invocation.getMethod();
-		Annotation[][] annotations = method.getParameterAnnotations();
-		Map<K, V> vars = new HashMap<K, V>();
-		int paramIndx = 0;
-		for (Annotation[] annPerParam : annotations) {
-			for (Annotation annotation : annPerParam) {
-				if (!annotationType.isAssignableFrom(annotation.getClass())) {
-					continue;
-				}
-				K pv = (K) annotation;
-				V v = (V) invocation.getArguments()[paramIndx];
-				vars.put(pv, v);
+  @SuppressWarnings("unchecked")
+  private <K extends Annotation, V> Map<K, V> mapOfAnnotationValues(Class<K> annotationType, MethodInvocation invocation) {
+    Method method = invocation.getMethod();
+    Annotation[][] annotations = method.getParameterAnnotations();
+    Map<K, V> vars = new HashMap<K, V>();
+    int paramIndx = 0;
+    for (Annotation[] annPerParam : annotations) {
+      for (Annotation annotation : annPerParam) {
+        if (!annotationType.isAssignableFrom(annotation.getClass())) {
+          continue;
+        }
+        K pv = (K) annotation;
+        V v = (V) invocation.getArguments()[paramIndx];
+        vars.put(pv, v);
 
-			}
-			paramIndx += 1;
-		}
-		return vars;
-	}
+      }
+      paramIndx += 1;
+    }
+    return vars;
+  }
 
 
-	/**
-	 * if there any arguments with the {@link org.camunda.bpm.engine.annotations.ProcessVariable} annotation,
-	 * then we feed those parameters into the business process
-	 *
-	 * @param invocation the invocation of the method as passed to the {@link org.aopalliance.intercept.MethodInterceptor#invoke(org.aopalliance.intercept.MethodInvocation)} method
-	 * @return returns the map of process variables extracted from the parameters
-	 * @throws Throwable thrown anything goes wrong
-	 */
-	protected Map<String, Object> processVariablesFromAnnotations(MethodInvocation invocation) throws Throwable {
+  /**
+   * if there any arguments with the {@link org.camunda.bpm.engine.annotations.ProcessVariable} annotation,
+   * then we feed those parameters into the business process
+   *
+   * @param invocation the invocation of the method as passed to the {@link org.aopalliance.intercept.MethodInterceptor#invoke(org.aopalliance.intercept.MethodInvocation)} method
+   * @return returns the map of process variables extracted from the parameters
+   * @throws Throwable thrown anything goes wrong
+   */
+  protected Map<String, Object> processVariablesFromAnnotations(MethodInvocation invocation) throws Throwable {
 
-		Map<ProcessVariable, Object> vars = this.mapOfAnnotationValues(ProcessVariable.class, invocation);
+    Map<ProcessVariable, Object> vars = this.mapOfAnnotationValues(ProcessVariable.class, invocation);
 
-		Map<String, Object> varNameToValueMap = new HashMap<String, Object>();
-		for (ProcessVariable processVariable : vars.keySet()) {
-			varNameToValueMap.put(processVariable.value(), vars.get(processVariable));
-		}
-		return varNameToValueMap;
+    Map<String, Object> varNameToValueMap = new HashMap<String, Object>();
+    for (ProcessVariable processVariable : vars.keySet()) {
+      varNameToValueMap.put(processVariable.value(), vars.get(processVariable));
+    }
+    return varNameToValueMap;
 
-	}
+  }
 }
