+++ engine-spring/core/src/main/java/org/camunda/bpm/engine/spring/components/aop/util/MetaAnnotationMethodMatcher.java
@@ -32,29 +32,29 @@ import java.lang.reflect.Method;
  */
 public class MetaAnnotationMethodMatcher extends AnnotationMethodMatcher {
 
-	private final Class<? extends Annotation> annotationType;
+  private final Class<? extends Annotation> annotationType;
 
 
-	/**
-	 * Create a new AnnotationClassFilter for the given annotation type.
-	 *
-	 * @param annotationType the annotation type to look for
-	 */
-	public MetaAnnotationMethodMatcher(Class<? extends Annotation> annotationType) {
-		super(annotationType);
-		this.annotationType = annotationType;
-	}
+  /**
+   * Create a new AnnotationClassFilter for the given annotation type.
+   *
+   * @param annotationType the annotation type to look for
+   */
+  public MetaAnnotationMethodMatcher(Class<? extends Annotation> annotationType) {
+    super(annotationType);
+    this.annotationType = annotationType;
+  }
 
 
-	@Override
-	@SuppressWarnings("rawtypes")
-	public boolean matches(Method method, Class targetClass) {
-		if (AnnotationUtils.getAnnotation(method, this.annotationType) != null) {
-			return true;
-		}
-		// The method may be on an interface, so let's check on the target class as well.
-		Method specificMethod = AopUtils.getMostSpecificMethod(method, targetClass);
-		return (specificMethod != method &&
-				(AnnotationUtils.getAnnotation(specificMethod, this.annotationType) != null));
-	}
+  @Override
+  @SuppressWarnings("rawtypes")
+  public boolean matches(Method method, Class targetClass) {
+    if (AnnotationUtils.getAnnotation(method, this.annotationType) != null) {
+      return true;
+    }
+    // The method may be on an interface, so let's check on the target class as well.
+    Method specificMethod = AopUtils.getMostSpecificMethod(method, targetClass);
+    return (specificMethod != method &&
+        (AnnotationUtils.getAnnotation(specificMethod, this.annotationType) != null));
+  }
 }
