+++ engine-spring/core/src/main/java/org/camunda/bpm/engine/spring/components/aop/util/Scopifier.java
@@ -30,58 +30,58 @@ import org.springframework.util.StringValueResolver;
  */
 public class Scopifier extends BeanDefinitionVisitor {
 
-	private final boolean proxyTargetClass;
+  private final boolean proxyTargetClass;
 
-	private final BeanDefinitionRegistry registry;
+  private final BeanDefinitionRegistry registry;
 
-	private final String scope;
+  private final String scope;
 
-	private final boolean scoped;
+  private final boolean scoped;
 
-	public static BeanDefinitionHolder createScopedProxy(String beanName, BeanDefinition definition, BeanDefinitionRegistry registry, boolean proxyTargetClass) {
-		BeanDefinitionHolder proxyHolder = ScopedProxyUtils.createScopedProxy(new BeanDefinitionHolder(definition, beanName), registry, proxyTargetClass);
-		registry.registerBeanDefinition(beanName, proxyHolder.getBeanDefinition());
-		return proxyHolder;
-	}
+  public static BeanDefinitionHolder createScopedProxy(String beanName, BeanDefinition definition, BeanDefinitionRegistry registry, boolean proxyTargetClass) {
+    BeanDefinitionHolder proxyHolder = ScopedProxyUtils.createScopedProxy(new BeanDefinitionHolder(definition, beanName), registry, proxyTargetClass);
+    registry.registerBeanDefinition(beanName, proxyHolder.getBeanDefinition());
+    return proxyHolder;
+  }
 
-	public Scopifier(BeanDefinitionRegistry registry, String scope, boolean proxyTargetClass, boolean scoped) {
-		super(new StringValueResolver() {
-			public String resolveStringValue(String value) {
-				return value;
-			}
-		});
-		this.registry = registry;
-		this.proxyTargetClass = proxyTargetClass;
-		this.scope = scope;
-		this.scoped = scoped;
-	}
+  public Scopifier(BeanDefinitionRegistry registry, String scope, boolean proxyTargetClass, boolean scoped) {
+    super(new StringValueResolver() {
+      public String resolveStringValue(String value) {
+        return value;
+      }
+    });
+    this.registry = registry;
+    this.proxyTargetClass = proxyTargetClass;
+    this.scope = scope;
+    this.scoped = scoped;
+  }
 
-	@Override
-	protected Object resolveValue(Object value) {
+  @Override
+  protected Object resolveValue(Object value) {
 
-		BeanDefinition definition = null;
-		String beanName = null;
-		if (value instanceof BeanDefinition) {
-			definition = (BeanDefinition) value;
-			beanName = BeanDefinitionReaderUtils.generateBeanName(definition, registry);
-		} else if (value instanceof BeanDefinitionHolder) {
-			BeanDefinitionHolder holder = (BeanDefinitionHolder) value;
-			definition = holder.getBeanDefinition();
-			beanName = holder.getBeanName();
-		}
+    BeanDefinition definition = null;
+    String beanName = null;
+    if (value instanceof BeanDefinition) {
+      definition = (BeanDefinition) value;
+      beanName = BeanDefinitionReaderUtils.generateBeanName(definition, registry);
+    } else if (value instanceof BeanDefinitionHolder) {
+      BeanDefinitionHolder holder = (BeanDefinitionHolder) value;
+      definition = holder.getBeanDefinition();
+      beanName = holder.getBeanName();
+    }
 
-		if (definition != null) {
-			boolean nestedScoped = scope.equals(definition.getScope());
-			boolean scopeChangeRequiresProxy = !scoped && nestedScoped;
-			if (scopeChangeRequiresProxy) {
-				// Exit here so that nested inner bean definitions are not
-				// analysed
-				return createScopedProxy(beanName, definition, registry, proxyTargetClass);
-			}
-		}
+    if (definition != null) {
+      boolean nestedScoped = scope.equals(definition.getScope());
+      boolean scopeChangeRequiresProxy = !scoped && nestedScoped;
+      if (scopeChangeRequiresProxy) {
+        // Exit here so that nested inner bean definitions are not
+        // analysed
+        return createScopedProxy(beanName, definition, registry, proxyTargetClass);
+      }
+    }
 
-		// Nested inner bean definitions are recursively analysed here
-		value = super.resolveValue(value);
-		return value;
-	}
+    // Nested inner bean definitions are recursively analysed here
+    value = super.resolveValue(value);
+    return value;
+  }
 }
