+++ engine-spring/core/src/main/java/org/camunda/bpm/engine/spring/components/config/xml/StateHandlerAnnotationBeanFactoryPostProcessor.java
@@ -39,50 +39,50 @@ import java.util.logging.Logger;
  */
 public class StateHandlerAnnotationBeanFactoryPostProcessor implements BeanFactoryPostProcessor {
 
-	private ProcessEngine processEngine ;
-	private Logger log = Logger.getLogger(getClass().getName());
+  private ProcessEngine processEngine ;
+  private Logger log = Logger.getLogger(getClass().getName());
 
-	public void setProcessEngine(ProcessEngine processEngine) {
-		this.processEngine = processEngine;
-	}
+  public void setProcessEngine(ProcessEngine processEngine) {
+    this.processEngine = processEngine;
+  }
 
-	private void configureDefaultActivitiRegistry(String registryBeanName, BeanDefinitionRegistry registry) {
-
-
-		if (!beanAlreadyConfigured(registry, registryBeanName, ActivitiStateHandlerRegistry.class)) {
-			String registryName =ActivitiStateHandlerRegistry.class.getName();
-			log.info( "registering a " + registryName + " instance under bean name "+ ActivitiContextUtils.ACTIVITI_REGISTRY_BEAN_NAME+ ".");
-
-			RootBeanDefinition rootBeanDefinition = new RootBeanDefinition();
-			rootBeanDefinition.setBeanClassName( registryName );
-			rootBeanDefinition.getPropertyValues().addPropertyValue("processEngine", this.processEngine);
-
-			BeanDefinitionHolder holder = new BeanDefinitionHolder(rootBeanDefinition, registryBeanName);
-			BeanDefinitionReaderUtils.registerBeanDefinition(holder, registry);
-		}
-	}
-
-	public void postProcessBeanFactory(ConfigurableListableBeanFactory beanFactory) throws BeansException {
-		if (beanFactory instanceof BeanDefinitionRegistry) {
-			BeanDefinitionRegistry registry = (BeanDefinitionRegistry) beanFactory;
-			configureDefaultActivitiRegistry(ActivitiContextUtils.ACTIVITI_REGISTRY_BEAN_NAME, registry);
+  private void configureDefaultActivitiRegistry(String registryBeanName, BeanDefinitionRegistry registry) {
 
 
-		} else {
-			log.info("BeanFactory is not a BeanDefinitionRegistry. The default '"
-					+ ActivitiContextUtils.ACTIVITI_REGISTRY_BEAN_NAME + "' cannot be configured.");
-		}
-	}
+    if (!beanAlreadyConfigured(registry, registryBeanName, ActivitiStateHandlerRegistry.class)) {
+      String registryName =ActivitiStateHandlerRegistry.class.getName();
+      log.info( "registering a " + registryName + " instance under bean name "+ ActivitiContextUtils.ACTIVITI_REGISTRY_BEAN_NAME+ ".");
 
-	private boolean beanAlreadyConfigured(BeanDefinitionRegistry registry, String beanName, Class clz) {
-		if (registry.isBeanNameInUse(beanName)) {
-			BeanDefinition bDef = registry.getBeanDefinition(beanName);
-			if (bDef.getBeanClassName().equals(clz.getName())) {
-				return true; // so the beans already registered, and of the right type. so we assume the user is overriding our configuration
-			} else {
-				throw new IllegalStateException("The bean name '" + beanName + "' is reserved.");
-			}
-		}
-		return false;
-	}
+      RootBeanDefinition rootBeanDefinition = new RootBeanDefinition();
+      rootBeanDefinition.setBeanClassName( registryName );
+      rootBeanDefinition.getPropertyValues().addPropertyValue("processEngine", this.processEngine);
+
+      BeanDefinitionHolder holder = new BeanDefinitionHolder(rootBeanDefinition, registryBeanName);
+      BeanDefinitionReaderUtils.registerBeanDefinition(holder, registry);
+    }
+  }
+
+  public void postProcessBeanFactory(ConfigurableListableBeanFactory beanFactory) throws BeansException {
+    if (beanFactory instanceof BeanDefinitionRegistry) {
+      BeanDefinitionRegistry registry = (BeanDefinitionRegistry) beanFactory;
+      configureDefaultActivitiRegistry(ActivitiContextUtils.ACTIVITI_REGISTRY_BEAN_NAME, registry);
+
+
+    } else {
+      log.info("BeanFactory is not a BeanDefinitionRegistry. The default '"
+          + ActivitiContextUtils.ACTIVITI_REGISTRY_BEAN_NAME + "' cannot be configured.");
+    }
+  }
+
+  private boolean beanAlreadyConfigured(BeanDefinitionRegistry registry, String beanName, Class clz) {
+    if (registry.isBeanNameInUse(beanName)) {
+      BeanDefinition bDef = registry.getBeanDefinition(beanName);
+      if (bDef.getBeanClassName().equals(clz.getName())) {
+        return true; // so the beans already registered, and of the right type. so we assume the user is overriding our configuration
+      } else {
+        throw new IllegalStateException("The bean name '" + beanName + "' is reserved.");
+      }
+    }
+    return false;
+  }
 }
