+++ engine-spring/core/src/main/java/org/camunda/bpm/engine/spring/impl/ApplicationContextElResolver.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.camunda.bpm.engine.spring;
+package org.camunda.bpm.engine.spring.impl;
 
 import java.beans.FeatureDescriptor;
 import java.util.Iterator;
