+++ engine-spring/core/src/main/java/org/camunda/bpm/engine/spring/impl/SpringExpressionManager.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.camunda.bpm.engine.spring;
+package org.camunda.bpm.engine.spring.impl;
 
 import java.util.Map;
 import org.camunda.bpm.engine.impl.el.ExpressionManager;
