+++ engine-spring/core/src/main/java/org/camunda/bpm/engine/spring/impl/SpringTransactionContextFactory.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.camunda.bpm.engine.spring;
+package org.camunda.bpm.engine.spring.impl;
 
 import org.camunda.bpm.engine.impl.cfg.TransactionContext;
 import org.camunda.bpm.engine.impl.cfg.TransactionContextFactory;
