+++ spring-boot-starter/starter/src/test/resources/application-nojpa.properties
@@ -1,4 +1,3 @@
-camunda.bpm.process-engine-name=testEngine
+camunda.bpm.process-engine-name=nojpaTestEngine
 camunda.bpm.database.schema-update=true
 camunda.bpm.history-level=audit
-camunda.bpm.dummy.enabled=false
